% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIDBH_module.R
\name{AIDBH_module}
\alias{AIDBH_module}
\title{Projects individual annual increment in DBH for Nothofagus from different parameters}
\usage{
AIDBH_module(
  SP = NA,
  DBH = NA,
  A = NA,
  SS = NA,
  SSCAL = NA,
  BALc = NA,
  ZONE = NA,
  AD = NA,
  SDI = NA,
  model = 1
)
}
\arguments{
\item{SP}{Specie (1:Rauli, 2:Roble, 3:Coihue, 4:Other)}

\item{DBH}{Diameter at breast height (cm) of the tree}

\item{A}{Breast height age (year) of the tree}

\item{SS}{Sociological status (1:dominant, 2:codominant, 3:intermediate, 4:suppressed)}

\item{SSCAL}{Calculated sociological status}

\item{BALc}{Basal area of larger trees for Nothofagus (m2/ha)}

\item{ZONE}{Growth zone of the plot (1, 2, 3, 4)}

\item{AD}{Dominant age (years) of the plot}

\item{SDI}{Stand Density Index (trees/ha) of the plot}

\item{model}{Model to use for annual DBH increment (1, 2, 3, 4) (default = 1)}
}
\value{
Individual annual increment in DBH (AIDBH, mm) (for projection)
}
\description{
Evaluates input variables to project individual annual increment in DBH for Nothofagus
based on tree-level and stand-level parameters. Projections are based in 1 year increments.
}
\examples{
# Example 1: Individual annual increment in DBH (AIDBH, mm) with model 1
Gest <- AIDBH_module(BALc=2, SDI=800, DBH=15, A=30, SS=1, model=1)
Gest

# Example 2: Individual annual increment in DBH (AIDBH, mm) with model 2
Gest <- AIDBH_module(BALc=2, SDI=800, DBH=15, A=30, SS=1, AD=50, SSCAL=0.5, model=2)
Gest

# Example 3: Individual annual increment in DBH (AIDBH, mm) with model 3
Gest <- AIDBH_module(SP=3, ZONE=2, BALc=2, SDI=800, DBH=15, A=30, SS=1, model=3)
Gest

# Example 4: Individual annual increment in DBH (AIDBH, mm) with model 4
Gest <- AIDBH_module(SP=3, ZONE=2, BALc=2, SDI=800, DBH=15, A=30, SS=1, AD=50, SSCAL=0.5, model=4)
Gest
}
\references{
Moreno et al. (2017). Individual-tree diameter growth models for mixed Nothofagus
second growth forests in southern Chile. Forests 8(12), 506.
}
\author{
P. Moreno and S.A. Gezan
}
