% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA99_module.R
\name{BA99module}
\alias{BA99module}
\title{Predicts and/or projects stand basal area for other species}
\usage{
BA99module(
  BA990 = NA,
  AD0 = NA,
  PNHAN0 = NA,
  PNHAN1 = NA,
  PBAN0 = NA,
  PBAN1 = NA,
  projection = FALSE
)
}
\arguments{
\item{BA990}{Basal area (m2/ha) of other species at current time (required for projections)}

\item{AD0}{Dominant age (years) of the stand at current time}

\item{PNHAN0}{Proportion of number of trees of Nothofagus (trees/ha) of the stand at time 0}

\item{PNHAN1}{Proportion of number of trees of Nothofagus (trees/ha) of the stand at time 1}

\item{PBAN0}{Proportion of basal area (m2/ha) of Nothofagus of the stand at time 0}

\item{PBAN1}{Proportion of basal area (m2/ha) of Nothofagus of the stand at time 1}

\item{projection}{if TRUE projection from BA0 is executed for a 1 year increment}
}
\value{
Basal area for other species (BA99, m2/ha) for the current age (for prediction)
or at AD0+1 (for projection)
}
\description{
Evaluates stand-level input variables to predict or project basal area for other species.
Projections are based in 1 year increments.
}
\examples{
# Example 1: Predicts Basal Area
BA99est <- BA99module(AD0=19, PNHAN0=0.81, PBAN0=0.91, projection=FALSE)
BA99est$BA990

# Example 2: Projects Basal Area
BA99est <- BA99module(BA990=3.2, AD0=19, PNHAN0=0.81, PNHAN1=0.81,
                      PBAN0=0.91, PBAN1=0.91, projection=TRUE)$BA991
BA99est
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
