% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA_module.R
\name{BAmodule}
\alias{BAmodule}
\title{Predicts and/or projects stand Basal Area (BA) from stand-level parameters}
\usage{
BAmodule(
  AD0 = NA,
  HD0 = NA,
  N0 = NA,
  BA0 = NA,
  HD1 = NA,
  N1 = NA,
  model = 1,
  projection = FALSE
)
}
\arguments{
\item{AD0}{Dominant age (years) of the stand at current time}

\item{HD0}{Dominant height (m) of the stand at current time}

\item{N0}{Number of trees (trees/ha) of the stand at current time}

\item{BA0}{Basal area (m2/ha) of the stand at current time (required for projections)}

\item{HD1}{Dominant height (m) of the stand at future projection time
(required for projections)}

\item{N1}{Number of trees (trees/ha) of the stand at future projection time
(required for projections)}

\item{model}{Model to use for estimation (1:non-linear fit, 2:linear fit) (default = 1)}

\item{projection}{if TRUE projection from BA0 is executed for a 1 year increment}
}
\value{
Basal area (BA0, m2/ha) for the current age (for prediction)
or at AD0+1 (for projection)
}
\description{
Evaluates stand-level input variables to predict and/or project basal area.
Projections are based in 1 year increments.
}
\examples{
# Example 1: Predicts basal area at a certain age AD
BAest <- BAmodule(AD0=20, HD0=17.2, N0=2730, model=1, projection=FALSE)
BAest$BA0

# Example 2: Projects basal area for different models
(BAest <- BAmodule(AD0=20, HD0=17.2, N0=2730, BA0=33.11, HD1=19.1,
                   N1=2610, model=1, projection=TRUE)$BA1)
(BAest <- BAmodule(AD0=20, HD0=17.2, N0=2730, BA0=33.11, HD1=19.1,
                   N1=2610, model=2, projection=TRUE)$BA1)
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
