% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAN_module.R
\name{BANmodule}
\alias{BANmodule}
\title{Predicts and/or projects stand basal area for Nothofagus (BAN) from stand level parameters}
\usage{
BANmodule(
  BAN0 = NA,
  AD0 = NA,
  SI = NA,
  NHA0 = NA,
  NHA1 = NA,
  PBAN0 = NA,
  PBAN1 = NA,
  projection = FALSE
)
}
\arguments{
\item{BAN0}{Basal area (m2/ha) of all Nothofagus at current time (required for projections)}

\item{AD0}{Dominant age (years) of the stand at time 0}

\item{SI}{Site Index (in m) at reference dominant age of 20 years of the stand}

\item{NHA0}{Number of trees (trees/ha) of the stand at time 0}

\item{NHA1}{Number of trees (trees/ha) of the stand at time 1}

\item{PBAN0}{Proportion of basal area (m2/ha) of Nothofagus of the stand at time 0}

\item{PBAN1}{Proportion of basal area (m2/ha) of Nothofagus of the stand at time 1}

\item{projection}{if TRUE projection from BA0 is executed for a 1 year increment}
}
\value{
Basal area for Nothofagus (BAN, m2/ha) for the current age (for prediction)
or at AD0+1 (for projection)
}
\description{
Evaluates stand-level input variables to predict and/or project basal area
for all Nothofagus species. Projections are based in 1 year increments.
}
\examples{
# Example 1: Predicts basal area
BAest <- BANmodule(AD0=19, SI=14, NHA0=2140, PBAN0 = 0.91, projection=FALSE)
BAest$BAN0

# Example 2: Projects basal area
BAest <- BANmodule(BAN0=36.5, AD0=19,  SI=14, NHA0=2730, NHA1=2650,
                   PBAN0=0.91, PBAN1=0.91, projection=TRUE)
BAest$BAN1
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
