% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_module.R
\name{comp_module}
\alias{comp_module}
\title{General compatibility module between tree-level and stand-level simulations}
\usage{
comp_module(sim.tree = NA, sim.stand = NA)
}
\arguments{
\item{sim.tree}{a simulated tree object from tree_simulator}

\item{sim.stand}{a simulated stand object from stand_simulator}
}
\value{
A list of elements and parameters with updated tables with compatibility. Outputs are the same as
        input_module object but these are updated for simulations.
}
\description{
General compatibility module that takes tree-level simulations and
makes a compatibility with the stand-level simulations. Two compatibility methods are
available PY: proportional compatibility, and PG: proportional growth compatibility.
The input object (tree or stand) needs to have the option requesting compatibility
(i.e, type='comp') together with the method: comptype ('PY' or 'PG') (default = 'PY')
}
\examples{
# Example 1: Proportional yield (PY)
# Tree information
tree.list <- plot_example
plot.tree <- input_module(ZONE=2, AD=28, HD=23.5, AF=45, type='tree', area=500,
                          tree.list=tree.list, ddiam=FALSE)
sim.tree <- core_module(input=plot.tree)
head(sim.tree$tree.list)
# Stand information from plot.tree$sp.table
BA <- c(1.086,38.915,0.0,0.313)
N <- c(60,780,0,80)
plot.stand <- input_module(ZONE=2, AD=28, HD=23.5, AF=45, N=N, BA=BA,
                           type='stand', ddiam=FALSE)
sim.stand <- core_module(input=plot.stand)
# Summary tree and stand by specie table
sim.tree$sp.table
sim.stand$sp.table
# Requesting Compatibility, for method 'PY'
sim.tree$comptype = 'PY'
sim.tree$type = 'comp'
sim.comp <- comp_module(sim.tree=sim.tree, sim.stand=sim.stand)
sim.comp$sp.table

# Example 2: Proportional growth (PG)
sim.tree$comptype = 'PG'
sim.tree$type = 'comp'
sim.comp <- comp_module(sim.tree=sim.tree, sim.stand=sim.stand)
sim.comp$sp.table
}
\author{
S. Palmas, S.A. Gezan and P. Moreno
}
