% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostic plots for regression fitted and predicted values}
\usage{
diagnostics(obs, pred)
}
\arguments{
\item{obs}{Array of observed values.}

\item{pred}{Array of predicted values.}
}
\value{
Four different diagnostic plots
}
\description{
Diagnostic plots for regression fitted and predicted values
}
\examples{
library(Nothopack)
fit <- lm(sr ~ ., data = LifeCycleSavings)
diagnostics(LifeCycleSavings$sr, predict(fit))

}

