% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diam_distr.R
\name{diam_dist}
\alias{diam_dist}
\title{Generates the diameter distribution for a given Nothofagus stand to each species/cohort}
\usage{
diam_dist(sp.table = NA, HD = NA, DOM.SP = NA, ZONE = NA, class = 5)
}
\arguments{
\item{sp.table}{table with stand-level information by specie and total. Columns are
SPECIES, N, BA, QD, and table is grouped by SPECIES (1:Rauli, 2:Roble, 3:Coigue, 4:Other, 0:Total).}

\item{HD}{Dominant height (m) of dominant specie in the current stand}

\item{DOM.SP}{Dominant specie (1: Rauli, 2: Roble, 3: Coigue, 4:Mixed)}

\item{ZONE}{Growth zone (1, 2, 3, 4)}

\item{class}{Diameter class width (cm) (default = 5)}
}
\value{
A 3-dimension matrix of values for each specie/cohort by diameter classes (on specified
increments) by SPECIES and then by all relevant columns.
}
\description{
It generates the diameter distribution for a given stand for each of the
species/cohorts using the method of parameter recovery. The diameter classes are based on
5 cm intervals (by default). Data can be read from \emph{input_module}.
}
\examples{
# Example: Generation of distribution from stand-level input data
BA <- c(36.5, 2.8, 0, 2.4)
N <- c(464, 23, 0, 48)
plot <- input_module(type='stand', ZONE=2, AD=28, AF=28, HD=23.5, N=N, BA=BA)
Dd <- diam_dist(sp.table=plot$sp.table, HD=plot$HD,
              DOM.SP=plot$DOM.SP, ZONE=plot$ZONE)
Dd[5,,]  # Total diameter distribution
# Ploting distribution for all species
barplot(as.matrix(Dd[5,,5]), main='Diameter Distribution all species',
        xlab='DBH Class', beside=TRUE, col=4)
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
