% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponent.R
\name{exponent}
\alias{exponent}
\title{Exponent function}
\usage{
exponent(a, pow)
}
\arguments{
\item{a}{parameter}

\item{pow}{parameter}
}
\value{
The remaining parameter
}
\description{
a^pow exponential function. Useful to avoid site index estimate errors
}
\examples{
library(Nothopack)
exponent(2, 2)
}
\seealso{
see \code{\link{remaining_calc_B}}
}

