% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_table.R
\name{fitness_table}
\alias{fitness_table}
\title{Fitness values for observed and fitted values.}
\usage{
fitness_table(obs, pred)
}
\arguments{
\item{obs}{Array of observed values.}

\item{pred}{Array of predicted values.}
}
\value{
A table with R^2 empirical, RMSE, and Bias%.
}
\description{
Fitness values for observed and fitted values.
}
\examples{
library(Nothopack)
fit <- lm(sr ~ ., data = LifeCycleSavings)
fitness_table(LifeCycleSavings$sr, predict(fit))
}

