% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domsp.R
\name{get_domsp}
\alias{get_domsp}
\title{Determines the dominant specie of a Nothofagus stand}
\usage{
get_domsp(BA = NA)
}
\arguments{
\item{BA}{Vector of basal area (m2/ha) of the stand (1: Rauli, 2: Roble, 3: Coigue, 4:Others)}
}
\value{
DOM.SP Dominant specie of the stand (1: Rauli, 2: Roble, 3: Coigue, 4:Mixed)
}
\description{
It determines the dominant Nothofagus specie of a given stand
based on stand-level basal area.
}
\examples{
# Example: Obtaining species from vector of Basal Area
BA <- c(4.5,2.8,41.6,2.4)
(DOM.SP <- get_domsp(BA))
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065, Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
