% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentile.R
\name{get_percentile}
\alias{get_percentile}
\title{Calculates percentiles for DBH based on vectors of DBH and N}
\usage{
get_percentile(DBH = NA, FT = NA, percentiles = c(0.15, 0.85))
}
\arguments{
\item{DBH}{Vector of diameter breast height values}

\item{FT}{Vector of expansion factor values (or N, in case of diameter distribution)}

\item{percentiles}{list of desired percentile values}
}
\value{
Data frame with desired DBH percentile values
}
\description{
If calculates percentiles for DBH based on the frequency
table provided by vectors of DBH and number of trees (N). Used for
expansion factors (or N) corrections.
}
\examples{
DBH <- c(36.5, 2.8, 0, 2.4)
N <- c(464, 23, 0, 48)
get_percentile(DBH=DBH, FT=N, percentiles=c(0.15, 0.85))
}
\author{
S. Palmas, S.A. Gezan and P. Moreno
}
