% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stand.R
\name{get_stand}
\alias{get_stand}
\title{Calculates missing stand-level variable from the set: N, BA, QD}
\usage{
get_stand(BA = NA, N = NA, QD = NA)
}
\arguments{
\item{BA}{Basal area (m2/ha) of the stand}

\item{N}{Number of trees (trees/ha) of the stand}

\item{QD}{Quadratic diameter (cm) of the stand}
}
\value{
The missing stand level parameter: BA, N or QD
}
\description{
When two of the stand level variables: basal area (BA, m2/ha),
number of trees (N, trees/ha) or quadratic diameter (QD, cm) are given, it
returns the value of the remaining missing stand-level parameter.
}
\examples{
# Example 1: Obtain quadratic diameter
(QD <- get_stand(BA=25.2, N=1400))

# Example 2: Obtain basal area
(BA <- get_stand(QD=17, N=1400))

# Example 3: Obtain number of trees
(N <- get_stand(BA=25.2, QD=17))
}
\author{
S. Palmas, S.A. Gezan and P. Moreno
}
