% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taper.R
\name{get_taper}
\alias{get_taper}
\title{Calculates stem diameter or height based on taper equation models}
\usage{
get_taper(
  SPECIES = NA,
  ZONE = NA,
  DBH = NA,
  HT = NA,
  di = NA,
  hi = NA,
  model = 2
)
}
\arguments{
\item{SPECIES}{Dominant specie (1:Rauli, 2:Roble, 3:Coigue, 4:Others or Mixed)}

\item{ZONE}{Growth zone of the corresponding stand}

\item{DBH}{Diameter at breast height (cm) of tree}

\item{HT}{Total tree height (m)}

\item{di}{Stem diameter (cm) at given stem height hi (m)}

\item{hi}{Stem height (m) at given stem diameter di (cm)}

\item{model}{Type of selected taper model (1:zone specific, 2:all zones) (default = 2)}
}
\value{
The missing component (di or hi) from the requested taper model. Or it returns the complete
taper on vectors d and h (evaluated each centimeter up to total height).
}
\description{
It reads input of tree-level information (SPECIE, DBH, HT), stand-level (dominant specie,
zone), and one taper component (di or hi) to calculate the other element (hi or di) based on
a selection of available taper equation models.
}
\examples{
# Example 1: Unknown diameter inside bark stem diameter
(di <- get_taper(SPECIES=1, ZONE=2, DBH=8.1, HT=10.2, hi=4.3, model=1)$di)

# Example 2: Unknown stem height
(hi <- get_taper(SPECIES=1, ZONE=2, DBH=8.1, HT=10.2, di=9.8, model=1)$hi)

# Example 3: Plotting complete taper model for a given tree
tree <- get_taper(SPECIES=1, ZONE=2, DBH=8.1, HT=10.2, di=9.8, model=1)
plot(tree$h,tree$d, ylab='di (cm)', xlab='hi (m)', type='l', col='blue')
}
\references{
Gezan, S.A. and Moreno, P. (2000). Informe Procedimientos y Resultados Modelos de Ahusamiento.
Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan and P. Moreno
}
