% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_stats.R
\name{gof_stats}
\alias{gof_stats}
\title{Calculates goodness-of-fit statistics}
\usage{
gof_stats(obs, pred)
}
\arguments{
\item{obs}{vector of observed values}

\item{pred}{vector of predicted values}
}
\value{
A table with goodness-of-fit statistics:
r2emp: empirical coefficient of correlation
RMSE (and RMSE%) root mean square error (raw and in percentage of mean of observed data)
BIAS (and BIAS%) bias (raw and in percentage of mean of observed data)
Theil's Inequality Coefficient (Ahlburg 1984)
}
\description{
It obtains goodness-of-fit statistics (r2emp, RMSE, BIAS, RMSE%, BIAS%, Theil)
using provided observed and fitted values. Missing values are removed from the list.
}
\examples{
x <- c(1:20)
yobs <- 10 + 0.1*x + rnorm(20, mean=0, sd=0.2)
fit <- lm(yobs ~ x)
gof_stats(yobs, predict(fit))
}
\references{
Alburhg, D.A. (1984). Forecast Evaluation and Improvement using Thiel's 
Decomposition. Journal of Forecasting 3(3):345-351
}
\author{
P. Moreno & S.A. Gezan
}
