% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_coef.R
\docType{data}
\name{hd_coef}
\alias{hd_coef}
\title{Dominant Height Coefficients for Nothofagus species}
\format{An object of class \code{data.frame} with 12 rows and 7 columns.}
\usage{
data(hd_coef)
}
\description{
Dominant height ecuation parameters:
HD = 0.3 + a [1 – {1 – (IS / a) c } ((E + 0.5) / 18)] 1/c
c = b0 + b1 * IS
}
\references{
Gezan and Ortega. 2001. Desarrollo de un simulador de rendimiento para renovales de Roble, Rauli y Coigue.
}
\keyword{datasets}

