% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hparam_coef.R
\docType{data}
\name{hparam_coef}
\alias{hparam_coef}
\title{Individual Parametric Height Coefficients for Nothofagus species}
\format{An object of class \code{data.frame} with 13 rows and 10 columns.}
\usage{
data(hparam_coef)
}
\description{
Parametric height ecuation parameters:
}
\details{
HT = b0 + b1 HD + b2 DC0.95 +
b3 exp(-0.08 DAP) + b4 HD3 exp(-0.08 DAP) + b5 DC3 exp(-0.08 DAP)
}
\references{
Moreno and Gezan. 2001. MODELOS PARAMETRIZADOS DE ALTURA PARA RENOVALES DE ROBLE, RAULI Y COIGUE
}
\keyword{datasets}

