% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHA_module.R
\name{NHAmodule}
\alias{NHAmodule}
\title{Projects number of trees (N) from stand-level parameters to the next year}
\usage{
NHAmodule(NHA0 = NA, QD0 = NA, DOM.SP = 4, model = 3, return.DQmax = FALSE)
}
\arguments{
\item{NHA0}{Number of trees (trees/ha) of the stand at current time 0}

\item{QD0}{Quadratic diameter (cm) of the stand at current time 0}

\item{DOM.SP}{Dominant specie (1:Rauli, 2:Roble, 3:Coigue, 4:Others or Mixed)}

\item{model}{Number of fitted model for N estimation (1:Original Reineke,
2:Re-fitted Reineke, 3:Reineke with correction factor) (default = 3)}

\item{return.DQmax}{If true returns DQ maximum from Reineke's model (default = FALSE)}
}
\value{
Number of trees (N1, m2/ha) for the next year (i.e. time 1)
}
\description{
Evaluates stand-level input variables to project number of trees per hectare
to the following year. This is a mortality model, as it does not incorporate ingrowth.
}
\examples{
# Example: Number of trees for next year with current NHA0
(N1 <- NHAmodule(NHA0=2730, QD0=12.43, model=1))
(N1 <- NHAmodule(NHA0=2730, QD0=12.43, model=2))
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile

Gezan, S.A., Ortega, A., Andenmatten, E. (2007). Diagramas de manejo de densidad para renovales
de roble, rauli y coigue en chile. Bosque (Valdivia) 28, 97-105
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
