% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nmodule.R
\name{Nmodule}
\alias{Nmodule}
\title{Projects number of trees (N) from stand level parameters to the next year}
\usage{
Nmodule(N0 = NA, QD0 = NA, model = 1)
}
\arguments{
\item{N0}{Number of trees (trees/ha) of the stand at current time.}

\item{QD0}{Quadratic Diameter (m2/ha) of the stand at current time.}

\item{model}{available fitted model to use for estimation.}
}
\value{
Number of trees (N1, m2/ha) for the next year (at EDOM0+1)
}
\description{
\code{Nmodule} Evaluates stand level input variables to project number of trees per hectare to the following year.
This is a mortality model, as it does not incorporate ingrowth.
}
\examples{
N1<-Nmodule(N0=2730,QD0=12.43,model=1)
N1
}
\author{
S.Gezan, S.Palmas and P.Moreno
}

