% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_module.R
\name{recruit_module}
\alias{recruit_module}
\title{Incorporates new young trees (ingrowth/recruitment) to a current stand's
diameter distribution (module in development)}
\usage{
recruit_module(sp.table = NA, HD = NA, DOM.SP = NA, ZONE = NA)
}
\arguments{
\item{sp.table}{table with stand-level information by specie and total, columns are
SPECIES, N, BA, QD, and table is grouped by SPECIES (1:Rauli, 2:Roble, 3:Coigue, 4:Other, 0:Total).}

\item{HD}{Dominant height (m) of dominant specie in the current stand}

\item{DOM.SP}{Dominant specie (1: Rauli, 2: Roble, 3: Coigue)}

\item{ZONE}{Growth zone (1, 2, 3, 4)}
}
\value{
A list with: Dd (data frame with updated diameter distribution including
recruits (N and BA are updated); Rec.SP (vector or random proportional assignation
of recruits to each specie; and sp.table (data frame with updated sp.table including
recuits.
}
\description{
It adds a 1% of the total number of trees (N, trees/ha), which is
assigned to the species proportional to their number of trees. These are then
assigned to the first diameter class and reported to be added.
In the future it should estimate the 'space' for new trees (high competition, low number of
new trees). The procedure generates a new diameter distribution and then add trees to small classes.
NOTE 1: This model is not fully implemented nor incorporated into the simulator.
NOTE 2: Adjustment of VTHA is not included here.
}
\examples{
# Example: Evaluating a new recruitment
BA <- c(36.5, 2.8, 1.6, 2.4)
N <- c(464, 23, 16, 48)
input <- input_module(ZONE=1, AD=28, HD=18.5, N=N, BA=BA, type='stand')
Dd <- diam_dist(sp.table=input$sp.table, HD=input$HD, DOM.SP=input$DOM.SP, ZONE=input$ZONE)
Dd.R <- recruit_module(sp.table=input$sp.table, HD=input$HD, DOM.SP=input$DOM.SP,
                       ZONE=input$ZONE)
Dd.R$Rec.SP
input$sp.table
Dd.R$sp.table
head(Dd[1,,])
head(Dd.R$Dd[1,,])
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S. Palmas, S.A. Gezan and P. Moreno
}
