% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_calc_A.R
\name{remaining_calc_A}
\alias{remaining_calc_A}
\title{Remaining Variable Calculator}
\usage{
remaining_calc_A(AB = NA, NHA = NA, DC = NA)
}
\arguments{
\item{NHA}{Number of trees per hectare.}

\item{DC}{Quadratic Diameter}

\item{BA}{Basal area in square meters.}
}
\value{
The remaining parameter
}
\description{
When two of Basal Area (AB), Number of trees per hectare (NHA) and quadratic Diamete (DC)
are given, it returns the value of the remaining parameter.
}
\examples{
library(Nothopack)
remaining_calc_A(AB = 25, NHA = 1400)
}

