% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_calc_B.R
\name{remaining_calc_B}
\alias{remaining_calc_B}
\title{Remaining Variable Calculator B}
\usage{
remaining_calc_B(dom_sp, zone, E = NA, HD = NA, IS = NA,
  hd.coef = hd_coef, ...)
}
\arguments{
\item{dom_sp}{Dominant species}

\item{zone}{Growth Zone}

\item{E}{Age in Years}

\item{HD}{dominant height}

\item{IS}{dite index}

\item{...}{additional arguments to be passed to the low level bisection function}
}
\value{
The remaining parameter
}
\description{
When two of Age (E), Dominant Height (HD) and Site Index (IS)
are given, it returns the value of the remaining parameter.
The coefficients for the Site index come from the Nothofagus Simulator
from Gezan.
}
\examples{
remaining_calc_B(dom_sp = 1, zone = 2, HD = 20, IS = 10)

}
\seealso{
\code{\link{hd_coef}}. For BA, DC and NHA remainine calc see \code{\link{remaining_calc_A}}
}

