% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_parameters.R
\name{stand_parameters}
\alias{stand_parameters}
\title{Inventory processing by calculating stand-level parameter based on individual tree data}
\usage{
stand_parameters(plotdata = NA, area = NA, extra = NA)
}
\arguments{
\item{plotdata}{Data frame with plot data including the columns: SPECIES (1, 2, 3, 4), DBH (cm) and HT (m)}

\item{area}{Area of plot (m2)}

\item{extra}{List of additional column numbers to process from plot data for summary on stand table}
}
\value{
A data frame with columns stand-table:
                                SPECIES (1:Rauli, 2:Roble, 3:Coigue, 4:Others, 0:All),
                                N (number of trees/ha), BA (basal area, m2/ha), and QD (quadratic diameter, cm).
        Also stand parameters:   DOM.SP (dominant specie), HD (dominant height, m),
                                PBAN (proportion of basal area of Nothofagus),
                                PNHAN (proportion of number of trees of Nothofagus), and
                                BAind (basal area for each individual tree, m2)
        If an extra list is provided, then additional summary columns are added to the
        stand-table grouped by specie.
}
\description{
Calculates all relevant stand-level parameters from an inventory plot for
each of the species and for all together (1:Rauli, 2:Roble, 3:Coigue, 4:Other, 0:All).
The input table should contain the colums of SPECIES (1:Rauli, 2:Roble, 3:Coigue, 4:Other)
together with their diameter at breast height (DBH, cm), total tree height (HT, m), and
expansion factor (FT = 10,000/area, ha/m2). If area is provided, then expansion factor is not
considered.
}
\examples{
# Example 1: Simple plot calculations
plotdata <- plot_example
head(plotdata,10)
plot.temp <- input_module(ZONE=1, AD=28, type='tree', area=500,
                          tree.list=plotdata, T_model=2)$tree.list
head(plot.temp,10)
plot.inv <- stand_parameters(plotdata=plot.temp, area=500)
plot.inv$DOM.SP
plot.inv$HD
plot.inv$sd

# Example 2: Plot calculations extended to additional columns
plotdata <- plot_example
head(plotdata,10)
plot.temp <- input_module(ZONE=1, AD=28, type='tree', area=500,
                          tree.list=plotdata, T_model=2)$tree.list
head(plot.temp,10)
plot.extra <- stand_parameters(plotdata=plot.temp, area=500, extra=8)
plot.extra$sd
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
