% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_randomizer.R
\name{stand_randomizer}
\alias{stand_randomizer}
\title{Simulated stand randomizer}
\usage{
stand_randomizer()
}
\value{
A table with tree information
}
\description{
It returns a randomized stand table with NA, ND, NO and XX species.
}
\examples{
stand_randomizer()
}

