% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_simulator.R
\name{stand_simulator}
\alias{stand_simulator}
\title{Simulates plot at the whole stand-level for the next period}
\usage{
stand_simulator(core.stand = NULL)
}
\arguments{
\item{core.stand}{An object originated from core_module}
}
\value{
A series of elements and parameters with updated tables. Outputs are the same as \emph{input_module}
        but these are updated for simulations. The main table is \emph{sp.table} that contains a stand table
        for the new simulations for the given parameters. Also, all other information is updated for
        time AF.
}
\description{
It simulates plot level growth, mortality (and recruitment) of a given stand
requiring stand-level parameters coming from core_module.
Simulations are done using stand-level models starting from intial age (AD0)
until final age (ADF) in increments of 1 year.
Note that recruitment is currently not considered.
}
\examples{
# Example: Input stand-level data to predict volume using stand-level functions to next year
BA <- c(36.5, 2.8, 1.6, 2.4)
N <- c(464, 23, 16, 48)
input <- input_module(ZONE=1, AD=28, HD=18.5, AF=29, N=N, BA=BA,
                      type='stand', ddiam=FALSE, V_model=1)
input$sp.table  # Current stand at AD (time 0)
current.plot <- stand_simulator(core.stand=input)
current.plot$sp.table # Stand at AF=AD+1
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
