% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_module.R
\name{thin_module}
\alias{thin_module}
\title{Executes thinning for a plot according to given stand- or tree-level information}
\usage{
thin_module(core.thin = NULL)
}
\arguments{
\item{core.thin}{input object generated from \emph{core_module}}
}
\value{
A series of elements and parameters with updated tables. Outputs are the same as \emph{input_module}
        but these are updated for simulations. The main table is \emph{sp.table} that contains a stand table
        for the new simulations for the given parameters. Also, all other information is updated for
        time AF.
}
\description{
It executes a thinning for a plot based on stand- or tree-level information.
The thinning goes across all species (Nothofagus and other species) using identical
criteria.
}
\examples{
# Example 1: Stand thinning (equal rate for every diameter class) at 30 years
#    from initial stand-level data with a residual basal area porcentual of 50\%
BA <- c(1.09, 38.92, 0, 0.31)
N <- c(60, 780, 0, 80)
stand.input <- input_module(ZONE=2, AD=28, AF=70, ATHIN=30, HD=18.4, N=N, BA=BA,
                            type='stand', ddiam=FALSE, BARp=50, QD_ba=1)
sims.thin <- core_module(input=stand.input)
sims.thin$data.sim
plot(sims.thin$data.sim$AGE, sims.thin$data.sim$BA, ylim=c(0,60),
     xlab='Age (years)', type='o', col='blue', ylab='BA (m2/ha)')
plot(sims.thin$data.sim$AGE, sims.thin$data.sim$VTOT, ylim=c(0,600),
     xlab='Age (years)', type='o', col='blue', ylab='VTOT (m3/ha)')

# Example 2: Simulation with tree thinning at 30 years starting from initial tree-level
#    data (or file). FT.thin is a random vector of thinning, 0=out, 1=in (to be added)
tree.list <- plot_example
head(tree.list)
FT.thin <- round(0.3+runif(length(tree.list$FT),0))
head(FT.thin, 20)
plot.input <- input_module(ZONE=1, AD=28, AF=40, type='tree', ATHIN=30, FT.thin=FT.thin,
                           area=500, tree.list=tree.list, T_model=2)
head(plot.input$tree.list, 10)
sim.thin <- core_module(input=plot.input)
head(sim.thin$tree.list, 10)
plot.input$sp.table
sim.thin$sp.table
sim.thin$data.sim
plot(sim.thin$data.sim$AGE, sim.thin$data.sim$BA, ylim=c(0,60),
     xlab='Age (years)', type='o', col='blue', ylab='BA (m2/ha)')
plot(sim.thin$data.sim$AGE, sim.thin$data.sim$VTOT, ylim=c(0,400),
     xlab='Age (years)', type='o', col='blue', ylab='VTOT (m3/ha)')
}
\author{
S.A. Gezan
}
