% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.ht.R
\name{tree.ht}
\alias{tree.ht}
\title{Estimates the missing total height of trees from a given plot}
\usage{
tree.ht(DBH, HT, DOM.SP = NA, ZONE = NA, HD = NA, QD = NA, method = 1)
}
\arguments{
\item{DBH}{Vector of diameter at breast height (cm)}

\item{HT}{Vector of total height (m) (with measured and/or missing values)}

\item{DOM.SP}{Dominant specie (1:Rauli, 2:Roble, 3:Coigue)}

\item{ZONE}{Growth zone (1, 2, 3, 4)}

\item{HD}{Dominant height (m) of dominant specie in the current stand}

\item{QD}{Quadratic diameter (cm) of the stand}

\item{method}{Method to use to estimate missing heights. 1:Parametrized DBH-height model
(which requires input: DBH, DOM.SP, ZONE, HD, QD), and 2:Fits a simple DBH-height model from
available measurements using the equation: \eqn{ln(HT) = b0 + b1/DBH} (default = 1)}
}
\value{
A list containing the following:
\itemize{
\item \code{HTFIN} A vector of final tree heigths (m), replacing the missing values for estimated heigths and
                   retaining the observed heights
\item \code{r2} A value with the coefficient of determination from the fitting the DBH-height model when method = 2
}
}
\description{
Estimates the total height of trees (m) that have missing height
that come from fitting a regression model on inventory data. For the missing
trees there are two methods: 1) to estimate tree heights according to a parametrized
height-DBH model, or 2) to estimate tree heights by fitting a simple height-DBH model
that requires at least 10 measurements. Missing values are indentified as 'NA', and
regression is fitted without making distinction of species.
All input vectors should correspond to the same tree ID.
}
\examples{
# Example 1: Method 1 - Parametrized DBH-height model
DBH <- c(9.3, 11.1, 15.5, 9, 14.8, 27.3, 11.4, 6.6, 12.6, 17.5,
         6.3, 7.2, 11.5, 13.6, 7.3, 12, 11.9, 8.1, 7.6, 5)
HT <- c(11.8, 12.3, NA, NA, 15.3, 18, 12, NA, 14.5, NA, NA,
        NA, NA, NA, 10.3, 14.6, NA, NA, NA, NA)
(ht.est <- tree.ht(DBH=DBH, HT=HT, DOM.SP=2, ZONE=2, HD=15, QD=12, method=1)$HTFIN)
plot(DBH, ht.est, ylab='HT (m)', xlab='DBH (cm)', xlim=c(0,30), ylim=c(0,20),
     type='p', col='blue')

# Example 2: Method 2 - Simple DBH-height model
DBH <- c(9.3, 11.1, 15.5, 9, 14.8, 27.3, 11.4, 6.6, 12.6, 17.5,
         6.3, 7.2, 11.5, 13.6, 7.3, 12, 11.9, 8.1, 7.6, 5)
HT <- c(11.8, 12.3, NA, NA, 15.3, 18, 12, NA, 14.5, NA, NA,
        NA, NA, NA, 10.3, 14.6, NA, NA, NA, NA)
(ht.est <- tree.ht(DBH=DBH, HT=HT, DOM.SP=2, ZONE=2, HD=15, QD=12, method=2)$HTFIN)
(r2 <- tree.ht(DBH=DBH, HT=HT, DOM.SP=2, ZONE=2, HD=15, QD=12, method=2)$r2)
plot(DBH, ht.est, ylab='HT (m)', xlab='DBH (cm)',
     xlim=c(0,30), ylim=c(0,20), type='p', col='darkblue')
}
\author{
P. Someda-Dias, and S.A. Gezan
}
