% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_covariates.R
\name{tree_covariates}
\alias{tree_covariates}
\title{Obtains/calculates stand- and tree-level covariates that feed into the individual DBH tree growth model}
\usage{
tree_covariates(ID = NA, FT = NA, SPECIES = NA, DBH = NA, ZONE = NA, SS = NA)
}
\arguments{
\item{ID}{Vector of identification of each tree}

\item{FT}{Vector of expansion factor (\emph{i.e.}, FT = 1/area) for each tree}

\item{SPECIES}{Vector of species code (1:Rauli, 2:Roble, 3:Coigue, 4:Other) for each tree}

\item{DBH}{Vector of diameter breast height (DBH, cm) for each tree}

\item{ZONE}{Growth zone (1, 2, 3, 4) for the plot}

\item{SS}{Vector of sociological status (range 1-4, where 1:Dominant, 2:Codominant, 3:Intermediate, 4:Supressed)}
}
\value{
A data frame containing the following vectors:
         ID (tree id), FT (expansion factor), DBH (diameter breast height, cm),
         ba (basal area tree, m2/ha), bac (basal area Nothofagus cohort, m2/ha),
         SPECIES (tree specie), ZONE (growth zone), SPZONA (concatenation specie-zone),
         N (number of trees/ha), BA (basal area, m2/ha), QD (quadratic diameter, cm),
         SDI (stand density index, trees/ha), SS (sociological status, range 1-4),
         BAL (basal area of larger trees, m2/ha), BALc (basal area of larger trees cohort Nothofagus, m2/ha),
         PSCAL (calculated sociological status as BAL/BA)
}
\description{
It obtains/calculates several stand- and tree-level covariables
that form part of the individual-tree DBH growth model (AIDBH). The input corresponds to tree data
originated from an inventory of a single plot, and they should all be on the same tree ID order.
Where there is missing data (e.g, SS) this information is completed.
}
\examples{
# Example 1: Simple data with SS to complete
ID <- c(1, 2, 3, 4)
FT <- c(400, 400, 400, 400)
SPECIES <- c(1, 1, 2, 4)
DBH <- c(15, 25, 37, 20)
SS <- c(NA, 2, NA, 2)
tree.fill <- tree_covariates(ID=ID, FT=FT, SPECIES=SPECIES, DBH=DBH, ZONE=1, SS=SS)
tree.fill

# Example 2: Complete inventory full plot
tree.list <- plot_example
tree.list$FT <- 10000/500
covar <- tree_covariates(ID=tree.list$ID, FT=tree.list$FT, SPECIES=tree.list$SPECIES,
                         DBH=tree.list$DBH, ZONE=2, SS=tree.list$SS)
head(covar, 10)
}
\references{
Moreno et al. (2017). Individual-tree diameter growth models for mixed Nothofagus
second growth forests in southern Chile. Forests 8(12), 506.
}
\author{
P. Moreno and S.A. Gezan
}
