% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_simulator.R
\name{tree_simulator}
\alias{tree_simulator}
\title{Simulates individual-tree growth in diameter (DBH) and mortality for next year}
\usage{
tree_simulator(core.tree = NULL)
}
\arguments{
\item{core.stand}{An object completely originated from core_module}
}
\value{
A series of elements and parameters with updated tables. Outputs are the same as input_module
        but these are updated for simulations. The main table is sp.table and tree.list that
        contain a stand table for the new simulations, and the tree list for the latest year, all
        based on the given parameters. Also, all other information is updated for time AF.
}
\description{
It simulates tree-level growth, mortality (and recruitment) of a given
stand requiring stand-level and tree-level parameters. Simulations are done using tree-level
models starting from intial age (AD0) until final age (ADF) in increments of 1 year.
The model of mortality is the same as whole-stand simulator.
Note, recruitment is currently not implemented.
}
\examples{
# Example: Simulation from Input from tree-level data (or file)
tree.list <- plot_example # use available data in example
head(tree.list)
plot <- input_module(ZONE=1, AD=28, AF=29, type='tree', area=500, tree.list=tree.list, T_model=2)
sims <- tree_simulator(core.tree=plot)
head(sims$tree.list)
plot$sp.table
sims$sp.table
plot$data.sim
sims$data.sim
}
\references{
Moreno, P.; Palmas, S.; Escobedo, F.; Cropper, W.; Gezan, S. Individual-tree diameter growth models
for mixed nothofagus second growth forests in southern chile. Forests 2017(8), 506
}
\author{
S.A. Gezan, S. Palmas and P. Moreno
}
