% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VOL_module.R
\name{Vmodule}
\alias{Vmodule}
\title{Estimates stand-level volume based on stand-level parameters}
\usage{
Vmodule(BA = NA, HD = NA, PNHAN = NA, PBAN = NA)
}
\arguments{
\item{BA}{Basal Area (m2/ha) of the current stand}

\item{HD}{Dominant height (m) of the current stand}

\item{PNHAN}{Proportion of Nothofagis trees in the stand (optional)}

\item{PBAN}{Proportion of Nothofagus basal area in the stand (optional)}
}
\value{
Total volume with bark (m3/ha) for the current conditions across all species
}
\description{
It reads stand-level input variables (BA, HD and PNHAN, PBAN) to estimate stand-level total
volume with bark (m3.ha). There are two models: 1:requires BA, HD, PNHAN and PBAN, and
2:requires BA, HD, PNHAN. The later model is more stable.
}
\examples{
# Example 1: Predicts volume with PNHAN and PBAN
Vmodule(BA=32, HD=15.9, PNHAN=0.81, PBAN=0.90)

# Example 2: Predicts volume with PNHAN
Vmodule(BA=32, HD=15.9, PNHAN=0.81)

# Example 3: Predicts volume
Vmodule(BA=32, HD=15.9)
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile
(updated version with recalculated stand volumes)
}
\author{
S.A. Gezan, P. Moreno, S. Palmas
}
