% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIND_module.R
\name{Vmodule_individual}
\alias{Vmodule_individual}
\title{Estimates individual-tree volume inside bark based on taper equations}
\usage{
Vmodule_individual(
  SPECIES = NA,
  ZONE = NA,
  DBH = NA,
  HT = NA,
  dmin = NA,
  blength = NA,
  stump = 0.3,
  model = 2
)
}
\arguments{
\item{SPECIES}{Species of the given tree (1:Rauli, 2:Roble, 3:Coigue)}

\item{ZONE}{Growth zone (1, 2, 3, 4) of the stand}

\item{DBH}{Diameter at breast height (cm)}

\item{HT}{Total tree height (m)}

\item{dmin}{Minimum stem diameter inside bark (cm)}

\item{blength}{Bole length (m)}

\item{stump}{Length of stump to discount (default = 0.3 m)}

\item{model}{Model selected for taper equation (1:zone specific, 2:all zones) (default = 2)}
}
\value{
Volume of stem section without bark (m3) based on the restrictions provided
}
\description{
It reads input of tree information (specie, DBH, HT), stand information (SPECIES, ZONE),
and of specification of restrictions (dmin, blength), to estimate inside bark volume (m3)
for the given tree based on different fitted taper equation models. Tree volume is calculated
starting on stump (0.3 m) until one of the following is reached: 1) stem diameter limit (dmin),
or 2) bole lenght (blength)
}
\examples{
# Example 1: Calculates tree volume for diameter limit of 5 cm (with stump of 0.3 m)
Vmodule_individual(SPECIES=1, ZONE=1, DBH=22.1, HT=18.2, dmin=5, model=1)

# Example 2: Calculates tree volume for a bole length of 6 m (with a stump of 0.3 m)
Vmodule_individual(SPECIES=1, ZONE=1, DBH=22.1, HT=18.2, blength=6, model=1)

# Example 3: Two ways to calculate total tree volume (with a stump of 0.3 m)
Vmodule_individual(SPECIES=1, ZONE=1, DBH=22.1, HT=18.2, dmin=0, model=1)
Vmodule_individual(SPECIES=1, ZONE=1, DBH=22.1, HT=18.2, blength=18.2, model=1)

# Example 4: Calculates total tree volume without discounting for stump
Vmodule_individual(SPECIES=1, ZONE=1, DBH=22.1, HT=18.2, dmin=0, stump=0, model=1)
}
\references{
Gezan, S.A. and Ortega, A. (2001). Desarrollo de un Simulador de Rendimiento para
Renovales de Roble, Rauli y Coigue. Reporte Interno. Projecto FONDEF D97I1065. Chile

Gezan, S.A. and Moreno, P. (2000b). INFORME MODELOS DE VOLUMEN.
Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan & P. Moreno
}
