% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPROD_module.R
\name{VPRODmodule}
\alias{VPRODmodule}
\title{Calculates tree volume (inside bark) based on taper equations for different products according to specifications}
\usage{
VPRODmodule(
  tree.list = NA,
  ZONE = NA,
  prod.list = NA,
  T_model = 2,
  stump = 0.3
)
}
\arguments{
\item{tree.list}{Data frame containing the following vectors:
ID (tree id), SPECIES (tree specie), DBH (diameter breast height, cm),
HT (total tree height, m). Additional columns can be included but these will be ignored}

\item{ZONE}{Growth zone of the corresponding stand}

\item{prod.list}{Data frame containing the following vectors:
Product (product abbreviation), Desc (description of product),
Dmin (minimum diameter for product, cm), Length (length of section, m)}

\item{T_model}{Model selected for taper equation (1:zone specific, 2:all zones) (default = 2)}

\item{stump}{Length of stump to discount (default = 0.3 m)}
}
\value{
Updated tree list with additional columns of individual tree
volume products (m3) (one column per product)
}
\description{
Individual tree volume for different products, according to a table of product specifications,
is calculated over all individuals provided in a tree list (based on DBH, HT, SPECIES).
A new tree list with additional columns is generated, which can be used to obtain a stand table.
Note: the table of product specificiacionts needs to be ordered from most to least important.
}
\examples{
# Example: Calculation of volumes for 5 different products with their stand table
tree.data <- plot_example
head(tree.data,10)  # Original data
tree.inv <- input_module(ZONE=1, AD=28, type='tree', area=500, tree.list=tree.data, T_model=2)
head(tree.inv$tree.list, 10)
Product <- c('SL', 'SS', 'PP', 'IF', 'RES')
Desc <- c('Saw-log (large)','Saw-log (short)','Panel','Industrial Firewood','Residual')
Dmin <- c(18, 16, 10, 5, 0)
Length <- c(4.1, 3.2, 2.44, 1.22, 99)
(Vprod <- data.frame(Product,Desc,Dmin,Length))
updated.tree.list <- VPRODmodule(tree.list=tree.inv$tree.list, ZONE=tree.inv$ZONE,
                               prod.list=Vprod, T_model=2)
head(updated.tree.list, 10)
tree.inv$sp.table
plot.extra <- stand_parameters(plotdata=updated.tree.list, area=500,
                  extra=c(8,9,10,11,12,13,14))
plot.extra$sd
}
\references{
Gezan, S.A. and Moreno, P. (2000b). Informe Modelos de Volumen.
Reporte Interno. Projecto FONDEF D97I1065. Chile
}
\author{
S.A. Gezan
}
