% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAmodule.R
\name{BAmodule}
\alias{BAmodule}
\title{Predicts and/or projects stand Basal Area (BA) from stand level parameters}
\usage{
BAmodule(EDOM0 = NA, HDOM0 = NA, N0 = NA, BA0 = NA, model = 1,
  projection = FALSE)
}
\arguments{
\item{EDOM0}{Dominant age (years) of the stand at current time.}

\item{N0}{Number of trees (trees/ha) of the stand at current time.}

\item{BA0}{Basal area (m2/ha) of the stand at curernt time (required for projections).}

\item{model}{available fitted model to use for estimation.}

\item{projection}{if TRUE projection from BA0 is executed for a 1 year increment.}

\item{HDOM}{Dominant height (m) of the stand at current time.}
}
\value{
Basal area (BA0, m2/ha) for the current age (for prediction)
or at EDOM0+1 (for projection)
}
\description{
\code{BAmodule} Evaluate stand level input variables to predict and/or project basal area.
Projections are based in year increments.
}
\examples{
# Example 1: Predicts Basal Area
BAest<-BAmodule(EDOM0=20,HDOM0=17.20,N0=2730,model=2,projection=FALSE)
BAest$BA0

# Example 2: Projects Basal Area
(BAest<-BAmodule(EDOM0=20,HDOM0=17.20,N0=2730,BA0=33.11,model=1,projection=TRUE)$BA1)
}
\author{
S.Gezan, S.Palmas and P.Moreno
}

