% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_height.R
\name{height_param}
\alias{height_param}
\title{Height Models in function of Whole Stand parameters}
\usage{
height_param(dom_sp, zone, HD = NA, DC = NA, DBH = NA,
  hparam.coef = hparam_coef, ...)
}
\arguments{
\item{HD}{Dominant height in meters.}

\item{DC}{Quadratic Diameter in centimeters.}

\item{DBH}{Diameter at Breast Height in centimeters.}
}
\value{
The individual Height in meters
}
\description{
Calculates individual tree heights in function of whole stand parameter
like Dominant Height (HD), Quadratic Diameter (DC) and also DBH
HT = b0 + b1 HD + b2 DC0.95 + b3 exp(-0.08 DAP) + b4 HD3 exp(0.08 DAP) + b5 DC3 exp(-0.08 DAP)
The coefficients come from the Nothofagus Simulator
from Gezan.
}
\examples{
height_param(HD = 15, DC = 12, DBH = 14)
library(Nothopack)
}

