% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Stand simulator path}
\usage{
simulator(stand, zone, EDOM0, area = 10000, sy = 30, ...)
}
\arguments{
\item{stand}{the stand to be simulated}

\item{zone}{zone of the stand}

\item{sy}{simuation years}

\item{age}{initial age of the stand}
}
\description{
\code{Nmodule} Simulates growth, mortality and recruitment of a given stand
}
\examples{
stand <- stand_randomizer()
simulator(stand, zone = 1, age = 15)
}
\author{
S.Gezan, S.Palmas and P.Moreno
}

