% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_parameters.R
\name{stand_parameters}
\alias{stand_parameters}
\title{Stand Parameter calculation}
\usage{
stand_parameters(stand, area, species.col = "ESPECIE", dbh.col = "DAP",
  height.col = "HT")
}
\arguments{
\item{stand:}{table. It must include the columns}

\item{area:}{Area of plot in square meters}

\item{species.col:}{name of species column name. Default to ESPECIES.}

\item{dbh.col:}{name of dbh column name. Default to DAP.}
}
\value{
A table with the columns BA (m2), NHA, DC, DAPM, MEDAD, HDOM, EDOM
}
\description{
When given a stand table it returns a table with stand parameters
}
\examples{
library(Nothopack)
stand_parameters(stand = stand, area = 2000)
}

